requirements.txt
streamlit
openai
faiss-cpu
pypdf
python-docx
python-dotenv
